Refactor AuthorityContext
